Add intermediate class to separate definitions
